rename Mypage to MyPage
