+++ src/main/java/kh/hobby1st/dto/MyPageDTO.java
@@ -1,5 +1,5 @@
 package kh.hobby1st.dto;
 
-public class MypageDTO {
+public class MyPageDTO {
 
 }
