+++ src/main/java/kh/hobby1st/service/MyPageService.java
@@ -1,32 +1,33 @@
-package kh.hobby1st.dao;
+package kh.hobby1st.service;
 
 import java.util.List;
 
-import org.mybatis.spring.SqlSessionTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Repository;
+import org.springframework.stereotype.Service;
 
+import kh.hobby1st.dao.MemberDAO;
+import kh.hobby1st.dao.MyPageDAO;
 import kh.hobby1st.dto.ClubListDTO;
 
-@Repository
-public class MypageDAO {
-	
+@Service
+public class MyPageService {
+
 	@Autowired
-	private SqlSessionTemplate mybatis;
-	
+	public MyPageDAO my_dao;
+
 	// 내가 만든 동호회 리스트
 	public List<ClubListDTO> clubList_make(String id) {
-		return mybatis.selectList("Mypage.clubList_make", id);
+		return my_dao.clubList_make(id);
 	}
-	
+
 	// 내가 가입한 동호회 리스트
 	public List<ClubListDTO> clubList_join(String id) {
-		return mybatis.selectList("Mypage.clubList_join", id);
+		return my_dao.clubList_join(id);
 	}
-	
+
 	// 내가 좋아요 한 동호회 리스트
 	public List<ClubListDTO> clubList_interest(String id) {
-		return mybatis.selectList("Mypage.clubList_interest", id);
+		return my_dao.clubList_interest(id);
 	}
 
 }
